Beta version 0.2.7! / Версия Beta 0.2.7! - added a more adequate formula for calculating the cost of pumping; - added a conversion script: -- allows you to represent large numbers in the form of short ones with signatures (mlm., sqdr., ...); - minor corrections; -------------------- - добавил более адекватную формулу расчёта стоимости прокачки; - добавил скрипт конвертации: -- позволяет представлять большие числа в виде коротких с подписями (млн., квдр. ...); - мелкие исправления;

+++ Assets/Scripts/Stats/StatsSystem.cs
@@ -44,11 +44,11 @@ namespace PlanetsColony
             }
         }
 
-        public float GetMaxShipsCount()
+        public uint GetMaxShipsCount()
         {
             return _maxShipsCount;
         }
-        public float GetActiveShipsCount()
+        public uint GetActiveShipsCount()
         {
             return _activeShipsCount;
         }
